backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
defusedxml
sqlalchemy
psycopg2-binary
python-jose[cryptography]
requests
python-dotenv
